bootloader: split pyi_utils.c
Split platform-specific implementations from `pyi_utils.c` into
`pyi_utils_posix.c` and `pyi_utils_win32.c`, to make them easier
to compare side-by-side (as well as reduce the amount of #ifdef
blocks).

The original `pyi_utils.c` file now contains implementations that
are common to all platforms/OSes; i.e., the parts that do not
require whole functions to be put under #ifdef blocks.
